// src/enums/popupEnum.ts
export enum PopupEnum {
    changePassword = 'changePassword',
    enterPassword = 'enterPassword',
    done = 'done',
    invoice = 'invoice',
    invoiceQr = 'invoiceQr',
    send = 'send',
    sendConfirm = 'sendConfirm',
    signConfirm = 'signConfirm',
    receive = 'receive',
    processing = 'processing',
    menuDropdown = 'menuDropdown',
    about = 'about',
    delete = 'delete',
    transaction = 'transaction',
    connectLedger = 'connectLedger',
    void = 'void',
}